tests/wpt/web-platform-tests/old-tests/submission/Infraware/Forms/contents/Forms/Not_Use/input_autocomplete.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>input_autocomplete</h3>
  </p>
  
  <li>input text in textbox and press enter key.</li>
  <li>When you input same text, you can see hint text.</li>
  
  <hr>

  <div id="log"></div>
  
  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action=""
      id="input_form">
  <p><input type='text' id='input_text'></p>
  </form>

  <script>
    
    var input_text = document.getElementById("input_text");
    input_text.autocomplete = 'on';

    if (typeof(input_text.autocomplete) == "string") {
      if (input_text.autocomplete != "on") {
        test(function() {
          assert_true(false, "autocomplete attribute is not correct.");
        });
      }
    } else {
      test(function() {
        assert_unreached("autocomplete attribute is not exist.");
      });
    }

  </script>

 </body>
</html>
